refactor(cli): give type-extractor a real extractTypes facade
Moves the two-pass scalar-bootstrap composition (previously
`extractTypesCore` in gen-orchestrator/orchestrator.ts) into
type-extractor/extract-types.ts as `extractTypes()`, the stage's real
entry point. The orchestrator's local `TypesResult` interface is
deleted in favor of the stage-owned `ExtractTypesResult` it now
returns; the pre-existing narrower `ExtractTypesResult` (produced by
`collectResults()`) is renamed to `CollectedTypesResult` to free up the
name for the facade's richer result type, since schema-generator only
ever consumed its `{ types, diagnostics }` shape.

orchestrator.ts, schema-generator/generate-schema.ts, and
schema-generator/integrator/result-integrator.ts now import
type-extractor exclusively through its index.ts facade, removing 8
deep imports into type-extractor internals. This collapses the
stage-boundary-type-extractor dependency-cruiser baseline from 25 to
14 entries (remaining entries belong to resolver-extractor's and
auto-type-generator's own facade work, out of scope here); no new
violations were introduced.
