+++ .dependency-cruiser-known-violations.json
@@ -263,83 +263,6 @@
   },
   {
     "type": "dependency",
-    "from": "packages/cli/src/gen-orchestrator/orchestrator.ts",
-    "to": "packages/cli/src/type-extractor/collector/result-collector.ts",
-    "unresolvedTo": "../type-extractor/collector/result-collector.js",
-    "dependencyTypes": ["local", "import"],
-    "rule": {
-      "severity": "error",
-      "name": "stage-boundary-type-extractor"
-    }
-  },
-  {
-    "type": "dependency",
-    "from": "packages/cli/src/gen-orchestrator/orchestrator.ts",
-    "to": "packages/cli/src/type-extractor/collector/scalar-collector.ts",
-    "unresolvedTo": "../type-extractor/collector/scalar-collector.js",
-    "dependencyTypes": ["local", "import"],
-    "rule": {
-      "severity": "error",
-      "name": "stage-boundary-type-extractor"
-    }
-  },
-  {
-    "type": "dependency",
-    "from": "packages/cli/src/gen-orchestrator/orchestrator.ts",
-    "to": "packages/cli/src/type-extractor/converter/graphql-converter.ts",
-    "unresolvedTo": "../type-extractor/converter/graphql-converter.js",
-    "dependencyTypes": ["local", "import"],
-    "rule": {
-      "severity": "error",
-      "name": "stage-boundary-type-extractor"
-    }
-  },
-  {
-    "type": "dependency",
-    "from": "packages/cli/src/gen-orchestrator/orchestrator.ts",
-    "to": "packages/cli/src/type-extractor/extractor/type-extractor.ts",
-    "unresolvedTo": "../type-extractor/extractor/type-extractor.js",
-    "dependencyTypes": ["local", "import"],
-    "rule": {
-      "severity": "error",
-      "name": "stage-boundary-type-extractor"
-    }
-  },
-  {
-    "type": "dependency",
-    "from": "packages/cli/src/gen-orchestrator/orchestrator.ts",
-    "to": "packages/cli/src/type-extractor/extractor/type-name-collector.ts",
-    "unresolvedTo": "../type-extractor/extractor/type-name-collector.js",
-    "dependencyTypes": ["local", "import"],
-    "rule": {
-      "severity": "error",
-      "name": "stage-boundary-type-extractor"
-    }
-  },
-  {
-    "type": "dependency",
-    "from": "packages/cli/src/gen-orchestrator/orchestrator.ts",
-    "to": "packages/cli/src/type-extractor/mapper/scalar-base-type-mapper.ts",
-    "unresolvedTo": "../type-extractor/mapper/scalar-base-type-mapper.js",
-    "dependencyTypes": ["local", "import"],
-    "rule": {
-      "severity": "error",
-      "name": "stage-boundary-type-extractor"
-    }
-  },
-  {
-    "type": "dependency",
-    "from": "packages/cli/src/gen-orchestrator/orchestrator.ts",
-    "to": "packages/cli/src/type-extractor/validator/type-validator.ts",
-    "unresolvedTo": "../type-extractor/validator/type-validator.js",
-    "dependencyTypes": ["local", "import"],
-    "rule": {
-      "severity": "error",
-      "name": "stage-boundary-type-extractor"
-    }
-  },
-  {
-    "type": "dependency",
     "from": "packages/cli/src/resolver-extractor/extractor/define-api-extractor.ts",
     "to": "packages/cli/src/type-extractor/extractor/field-type-resolver.ts",
     "unresolvedTo": "../../type-extractor/extractor/field-type-resolver.js",
@@ -377,50 +300,6 @@
     "to": "packages/cli/src/type-extractor/collector/scalar-collector.ts",
     "unresolvedTo": "../../type-extractor/collector/scalar-collector.js",
     "dependencyTypes": ["local", "type-only", "import"],
-    "rule": {
-      "severity": "error",
-      "name": "stage-boundary-type-extractor"
-    }
-  },
-  {
-    "type": "dependency",
-    "from": "packages/cli/src/schema-generator/generate-schema.ts",
-    "to": "packages/cli/src/type-extractor/collector/scalar-collector.ts",
-    "unresolvedTo": "../type-extractor/collector/scalar-collector.js",
-    "dependencyTypes": ["local", "type-only", "import"],
-    "rule": {
-      "severity": "error",
-      "name": "stage-boundary-type-extractor"
-    }
-  },
-  {
-    "type": "dependency",
-    "from": "packages/cli/src/schema-generator/generate-schema.ts",
-    "to": "packages/cli/src/type-extractor/converter/graphql-converter.ts",
-    "unresolvedTo": "../type-extractor/converter/graphql-converter.js",
-    "dependencyTypes": ["local", "import"],
-    "rule": {
-      "severity": "error",
-      "name": "stage-boundary-type-extractor"
-    }
-  },
-  {
-    "type": "dependency",
-    "from": "packages/cli/src/schema-generator/generate-schema.ts",
-    "to": "packages/cli/src/type-extractor/types/index.ts",
-    "unresolvedTo": "../type-extractor/types/index.js",
-    "dependencyTypes": ["local", "type-only", "import"],
-    "rule": {
-      "severity": "error",
-      "name": "stage-boundary-type-extractor"
-    }
-  },
-  {
-    "type": "dependency",
-    "from": "packages/cli/src/schema-generator/integrator/result-integrator.ts",
-    "to": "packages/cli/src/type-extractor/collector/scalar-collector.ts",
-    "unresolvedTo": "../../type-extractor/collector/scalar-collector.js",
-    "dependencyTypes": ["local", "import"],
     "rule": {
       "severity": "error",
       "name": "stage-boundary-type-extractor"
