+++ packages/cli/src/gen-orchestrator/orchestrator.ts
@@ -31,25 +31,14 @@ import {
   extractDirectiveDefinitions,
   toPosixPath,
 } from "../shared/index.js";
-import { collectResults } from "../type-extractor/collector/result-collector.js";
 import {
-  type CollectedScalarType,
   type ConfigScalarMapping,
-  collectScalars,
-  type ScalarMetadataInfo,
-} from "../type-extractor/collector/scalar-collector.js";
-import { convertToGraphQL } from "../type-extractor/converter/graphql-converter.js";
-import {
-  extractTypesFromProgram,
+  collectDeclaredTypeNames,
+  type ExtractTypesResult,
+  extractTypes,
   type GlobalTypeMapping,
-} from "../type-extractor/extractor/type-extractor.js";
-import { collectDeclaredTypeNames } from "../type-extractor/extractor/type-name-collector.js";
-import type { ExtractedTypeInfo } from "../type-extractor/index.js";
-import {
-  buildScalarMappingTable,
   type ScalarBaseTypeMappingTable,
-} from "../type-extractor/mapper/scalar-base-type-mapper.js";
-import { validateTypes } from "../type-extractor/validator/type-validator.js";
+} from "../type-extractor/index.js";
 import { scanDirectory } from "./infra/file-scanner.js";
 import { createSharedProgram } from "./infra/program-factory.js";
 import { writeFiles } from "./writer/file-writer.js";
@@ -76,17 +65,6 @@ export interface GenerationResult {
   readonly diagnostics: ReadonlyArray<Diagnostic>;
 }
 
-interface TypesResult {
-  types: ReturnType<typeof collectResults>["types"];
-  extractedTypes: ReadonlyArray<ExtractedTypeInfo>;
-  diagnostics: Diagnostics;
-  detectedScalarNames: ReadonlyArray<string>;
-  detectedScalars: ReadonlyArray<ScalarMetadataInfo>;
-  collectedScalars: ReadonlyArray<CollectedScalarType>;
-  scalarMappingTable: ScalarBaseTypeMappingTable | null;
-  discoveredTypeNames: ReadonlySet<string>;
-}
-
 interface ResolversResult {
   queryFields: { fields: ReadonlyArray<GraphQLFieldDefinition> };
   mutationFields: { fields: ReadonlyArray<GraphQLFieldDefinition> };
@@ -109,7 +87,7 @@ interface PipelineContext {
   readonly knownTypeNames: ReadonlySet<string> | null;
   readonly knownTypeSymbols: ReadonlyMap<string, ts.Symbol> | null;
   readonly underlyingSymbolToTypeName: ReadonlyMap<ts.Symbol, string> | null;
-  readonly typesResult: TypesResult | null;
+  readonly typesResult: ExtractTypesResult | null;
   readonly resolversResult: ResolversResult | null;
   readonly directiveDefinitions: DirectiveDefinitionInfo[] | null;
   readonly scalarConfig: ScalarConfig | null;
@@ -137,125 +115,6 @@ function hasErrors(
     typesResult.diagnostics.errors.length > 0 ||
     resolversResult.diagnostics.errors.length > 0
   );
-}
-
-interface ExtractTypesCoreParams {
-  readonly program: ts.Program;
-  readonly sourceFiles: ReadonlyArray<string>;
-  readonly customScalarNames: ReadonlyArray<string>;
-  readonly globalTypeMappings: ReadonlyArray<GlobalTypeMapping>;
-  readonly configScalars: ReadonlyArray<ConfigScalarMapping>;
-  readonly sourceRoot: string | null;
-  readonly knownTypeNames: ReadonlySet<string>;
-  readonly knownTypeSymbols: ReadonlyMap<string, ts.Symbol>;
-  readonly underlyingSymbolToTypeName: ReadonlyMap<ts.Symbol, string>;
-}
-
-function extractTypesCore(params: ExtractTypesCoreParams): TypesResult {
-  const {
-    program,
-    sourceFiles,
-    customScalarNames,
-    globalTypeMappings,
-    configScalars,
-    sourceRoot,
-    knownTypeNames,
-    knownTypeSymbols,
-    underlyingSymbolToTypeName,
-  } = params;
-  const allDiagnostics: Diagnostic[] = [];
-
-  // Pass 1: Extract types and detect scalars (without scalar mapping table)
-  const pass1Result = extractTypesFromProgram(program, sourceFiles, {
-    globalTypeMappings,
-    knownTypeNames,
-    knownTypeSymbols,
-    underlyingSymbolToTypeName,
-    scalarMappingTable: null,
-  });
-
-  // Build scalar mapping table from detected scalars
-  const scalarMappingTable =
-    pass1Result.detectedScalars.length > 0
-      ? buildScalarMappingTable({
-          detectedScalars: pass1Result.detectedScalars,
-          checker: program.getTypeChecker(),
-          program,
-        })
-      : null;
-
-  // Report scalar mapping conflicts as diagnostics
-  if (scalarMappingTable) {
-    for (const [, conflict] of scalarMappingTable.conflicts) {
-      const baseTypeName = conflict.baseTypeSymbol.getName();
-      const scalarNames = conflict.conflictingScalars
-        .map((s) => s.scalarName)
-        .join(", ");
-      allDiagnostics.push({
-        code: conflict.code,
-        message: `Base type '${baseTypeName}' maps to multiple scalars: ${scalarNames}. Use explicit scalar type instead of the base type.`,
-        severity: "error",
-        location: null,
-      });
-    }
-  }
-
-  // Pass 2: Re-extract types with scalar mapping table (if we have detected scalars)
-  const extractionResult = scalarMappingTable
-    ? extractTypesFromProgram(program, sourceFiles, {
-        globalTypeMappings,
-        knownTypeNames,
-        knownTypeSymbols,
-        underlyingSymbolToTypeName,
-        scalarMappingTable,
-      })
-    : pass1Result;
-
-  allDiagnostics.push(...extractionResult.diagnostics);
-
-  const allCustomScalarNames = [
-    ...customScalarNames,
-    ...extractionResult.detectedScalarNames,
-  ];
-
-  const scalarValidationResult = collectScalars(
-    extractionResult.detectedScalars,
-    configScalars,
-    { sourceRoot },
-  );
-  const collectedScalars: CollectedScalarType[] = scalarValidationResult.success
-    ? [...scalarValidationResult.data]
-    : [];
-  if (!scalarValidationResult.success) {
-    for (const error of scalarValidationResult.errors) {
-      allDiagnostics.push({
-        code: error.code,
-        message: error.message,
-        severity: error.severity,
-        location: null,
-      });
-    }
-  }
-
-  const conversionResult = convertToGraphQL(extractionResult.types);
-  allDiagnostics.push(...conversionResult.diagnostics);
-
-  const validationResult = validateTypes({
-    types: conversionResult.types,
-    customScalarNames: allCustomScalarNames,
-  });
-  allDiagnostics.push(...validationResult.diagnostics);
-
-  const result = collectResults(conversionResult.types, allDiagnostics);
-  return {
-    ...result,
-    extractedTypes: extractionResult.types,
-    detectedScalarNames: extractionResult.detectedScalarNames,
-    detectedScalars: extractionResult.detectedScalars,
-    collectedScalars,
-    scalarMappingTable,
-    discoveredTypeNames: extractionResult.discoveredTypeNames,
-  };
 }
 
 function convertArgsToInputValues(
@@ -621,7 +480,7 @@ function extractTypesStep(ctx: PipelineContext): PipelineContext {
   const { customScalarNames, globalTypeMappings, configScalars } =
     ctx.scalarConfig;
 
-  const typesResult = extractTypesCore({
+  const typesResult = extractTypes({
     program: ctx.program,
     sourceFiles: ctx.sourceFiles,
     customScalarNames,
