+++ packages/cli/src/type-extractor/index.ts
@@ -4,12 +4,25 @@ export type {
   GraphQLTypeInfo,
   GraphQLTypeKind,
 } from "../core/index.js";
-export type {
-  CollectedScalarType,
-  ConfigScalarMapping,
-  DescriptionSource,
-  ScalarMetadataInfo,
+export type { CollectedTypesResult } from "./collector/result-collector.js";
+export {
+  type CollectedScalarType,
+  type ConfigScalarMapping,
+  type DescriptionSource,
+  mergeDescriptions,
+  type ScalarMetadataInfo,
 } from "./collector/scalar-collector.js";
-export type { ExtractTypesResult } from "./extract-types.js";
+export { convertToGraphQL } from "./converter/graphql-converter.js";
+export {
+  type ExtractTypesParams,
+  type ExtractTypesResult,
+  extractTypes,
+} from "./extract-types.js";
+export type { GlobalTypeMapping } from "./extractor/type-extractor.js";
+export {
+  collectDeclaredTypeNames,
+  type TypeNameCollectionResult,
+} from "./extractor/type-name-collector.js";
+export type { ScalarBaseTypeMappingTable } from "./mapper/scalar-base-type-mapper.js";
 
 export type { ExtractedTypeInfo } from "./types/index.js";
